pyproject.toml
[tool.poetry]
name = "homebot"
version = "1.0.0"
description = "A Telegram bot written in Python"
authors = ["Sebastiano Barezzi <barezzisebastiano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.0"
python-dotenv = "^0.14.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
